refactor row content code to allow reordering for US
